+++ google_sign_in_all_platforms_interface/lib/src/extensions/platform.dart
@@ -1,8 +1,3 @@
-import 'dart:io';
-
-///Whether the current [Platform] is a desktop or not.
-bool get isDesktop =>
-    Platform.isMacOS || Platform.isLinux || Platform.isWindows;
-
-///Whether the current [Platform] is a mobile or not.
-bool get isMobile => Platform.isIOS || Platform.isWindows;
+export 'platform_stub.dart'
+    if (dart.library.html) 'platform_web.dart'
+    if (dart.library.io) 'platform_io.dart';