+++ google_sign_in_all_platforms_interface/lib/src/method_channel_google_sign_in_all_platforms.dart
@@ -1,15 +1,15 @@
-import 'package:flutter/foundation.dart' show visibleForTesting;
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_sign_in_all_platforms_interface/google_sign_in_all_platforms_interface.dart';
-import 'package:google_sign_in_all_platforms_interface/src/credentials.dart';
-import 'package:http/src/client.dart';
+import 'package:http/http.dart';
 
-/// An implementation of [GoogleSignInAllPlatformsInterface] that uses method channels.
-class MethodChannelGoogleSignInAllPlatforms extends GoogleSignInAllPlatformsInterface {
+/// An implementation of [GoogleSignInAllPlatformsInterface] that uses method
+/// channels.
+class MethodChannelGoogleSignInAllPlatforms
+    extends GoogleSignInAllPlatformsInterface {
   /// The method channel used to interact with the native platform.dart.
   @visibleForTesting
   final methodChannel = const MethodChannel('google_sign_in_all_platforms');
-
 
   @override
   Future<GoogleSignInCredentials?> signInOffline() {
@@ -32,6 +32,12 @@ class MethodChannelGoogleSignInAllPlatforms extends GoogleSignInAllPlatformsInte
   @override
   Future<Client?> getAuthenticatedClient() {
     // TODO(vishnuagbly): implement getAuthenticatedClient
+    throw UnimplementedError();
+  }
+
+  @override
+  Widget? signInButton({GSIAPButtonConfig? config}) {
+    // TODO: implement signInButton
     throw UnimplementedError();
   }
 }